requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-environ==0.11.2
psycopg2-binary==2.9.3
psycopg2==2.9.3
gevent==21.8.0
whitenoise==5.3.0
gunicorn==20.1.0
django-cors-headers==3.8.0
django-extensions==3.1.3
django-filter==2.4.0
django-silk==4.3.0
django-storages==1.11.1
drf-spectacular==0.18.1
pre-commit==2.15.0
PyJWT==1.7.1
requests==2.26.0
bandit==1.7.0
autoflake==2.3.1
black==24.4.0
flake8==7.0.0
pep8-naming==0.12.1
isort==5.13.2
djangorestframework-simplejwt
drf-spectacular==0.18.1
boto3==1.18.32
